pyproject.toml
[project]
name = "toddle-ops"
version = "0.3.0"
description = "Fast and Easy Project Generation for Exhausted Caregivers"
readme = "README.md"
authors = [
    { name = "asenetcky", email = "me@asenetcky.dev" }
]
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncio>=4.0.0",
    "gcloud>=0.18.3",
    "google-adk>=1.19.0",
    "mcp>=1.22.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.4.1",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
    "supabase>=2.5.3",
]

[project.scripts]
toddle-ops = "toddle_ops:main"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[project.optional-dependencies]
dev = [
    "google-adk[eval]>=1.19.0",
    "marimo>=0.18.0",
]
test = [
    "pytest>=8.4.0",
]

[dependency-groups]
dev = [
    "google-adk[eval]>=1.19.0",
    "marimo>=0.18.0",
    "opentelemetry-instrumentation-google-genai>=0.4b0",
]
test = [
    "pytest>=8.4.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
